+++ linlearn/datasets/_bank.py
@@ -41,21 +41,21 @@ def load_bank(download_if_missing=True):
     data_path = join(data_dir, "bank.zip")
 
     dtype = {
-        "age": np.int,
+        "age": int,
         "job": "category",
         "marital": "category",
         "education": "category",
         "default": "category",
-        "balance": np.int,
+        "balance": int,
         "housing": "category",
         "loan": "category",
         "contact": "category",
         "day": "category",
         "month": "category",
-        "duration": np.int,
-        "campaign": np.int,
-        "pdays": np.int,
-        "previous": np.int,
+        "duration": int,
+        "campaign": int,
+        "pdays": int,
+        "previous": int,
         "poutcome": "category",
     }
     dataset = Dataset.from_dtype(
